requirement.txt
Appium-Python-Client==0.50
subprocess
pyautogui==0.9.1
pandas==2.2.1
unittest
